+++ packages/create-block/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.8.2 (2020-02-26)
 
-- Fixed buggy check for minimum system requirements when run with `npx` and `npm init`.
+- Fixed buggy check for minimum system requirements when run with `npx` and `npm init` ([#20456](https://github.com/WordPress/gutenberg/pull/20456)).
 
 ## 0.8.1 (2020-02-25)